requirements.txt
# mysql-connector-python==8.4.0
django==4.2.24
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1